+++ KioChess/Engine/Sorting/Sorters/MoveSorter.cs
@@ -37,7 +37,7 @@ public abstract class MoveSorter<T>:MoveSorterBase where T:AttackCollection
     internal override void ProcessCaptureMove(AttackBase attack)
     {
         attack.Captured = Board.GetPiece(attack.To);
-        short attackValue = Board.StaticExchange(attack);
+        int attackValue = Board.StaticExchange(attack);
         if (attackValue > 0)
         {
             attack.See = attackValue;
@@ -74,7 +74,7 @@ public abstract class MoveSorter<T>:MoveSorterBase where T:AttackCollection
         var attack = promotions[0];
         attack.Captured = Board.GetPiece(attack.To);
 
-        short attackValue = Board.StaticExchange(attack);
+        int attackValue = Board.StaticExchange(attack);
         if (attackValue > 0)
         {
             AttackCollection.AddWinCaptures(promotions,attackValue);
@@ -132,12 +132,12 @@ public abstract class MoveSorter<T>:MoveSorterBase where T:AttackCollection
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void WhitePromotion(PromotionList moves)
     {
-        short max = short.MinValue;
+        int max = short.MinValue;
         for (byte i = 0; i < attackList.Count; i++)
         {
             var attack = attackList[i];
             attack.Captured = WhitePawn;
-            short see = Board.StaticExchange(attack);
+            int see = Board.StaticExchange(attack);
             if (see > max)
             {
                 max = see;
@@ -157,12 +157,12 @@ public abstract class MoveSorter<T>:MoveSorterBase where T:AttackCollection
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void BlackPromotion(PromotionList moves)
     {
-        short max = short.MinValue;
+        int max = short.MinValue;
         for (byte i = 0; i < attackList.Count; i++)
         {
             var attack = attackList[i];
             attack.Captured = BlackPawn;
-            short see = Board.StaticExchange(attack);
+            int see = Board.StaticExchange(attack);
             if (see > max)
             {
                 max = see;
